feat: Splits HawtioLogin to allow for custom form #594
* HawtioLogin.tsx
 * Looks at HawtConfig for title
 * Checks if there is an active login plugin available and substitutes its
   component for the login form
 * Should no plugin be available then the default HawtioLoginForm is loaded
 * Awaits on all plugins being loaded since otherwise the plugins array
   from usePlugins() will be empty

* HawtioLoginForm.tsx
 * Default delegate form that a user populate with user/password

* core.ts
 * Adds isLogin property to the Plugin interface to flag the plugin as a
   login plugin

* HawtioPage.tsx
 * Ensures that the login plugins are filtered out of the page plugins
